refactor: mock_broker_adapter
