+++ tests/autotrader/live_executor/conftest.py
@@ -12,7 +12,7 @@ to ensure complete isolation between tests.
 
 import pytest
 
-from python.framework.testing.mock_adapter import MockBrokerAdapter, MockExecutionMode
+from python.framework.testing.mock_broker_adapter import MockBrokerAdapter, MockExecutionMode
 from python.framework.testing.mock_order_execution import MockOrderExecution
 from python.framework.trading_env.live.live_trade_executor import LiveTradeExecutor
 from python.framework.trading_env.live.live_request_processor import LiveRequestProcessor
