+++ tests/autotrader/live_executor/test_broker_trade_records.py
@@ -16,7 +16,7 @@ that #327 Drift Audit consumes.
 
 from datetime import datetime, timezone
 
-from python.framework.testing.mock_adapter import MockBrokerAdapter, MockExecutionMode
+from python.framework.testing.mock_broker_adapter import MockBrokerAdapter, MockExecutionMode
 from python.framework.testing.mock_order_execution import MockOrderExecution
 from python.framework.types.live_types.live_request_types import (
     TradesQueryJob,
