+++ tests/autotrader/live_executor/test_live_executor_mock.py
@@ -8,7 +8,7 @@ Each test creates a fresh executor via conftest fixtures.
 MockOrderExecution provides tick feeding for pending order processing.
 """
 
-from python.framework.testing.mock_adapter import MockExecutionMode
+from python.framework.testing.mock_broker_adapter import MockExecutionMode
 from python.framework.testing.mock_order_execution import MockOrderExecution
 from python.framework.types.trading_env_types.order_types import (
     OrderType,
@@ -283,7 +283,7 @@ class TestNotLiveCapable:
     def test_non_live_adapter_raises(self):
         """LiveTradeExecutor raises ValueError for non-live adapter."""
         from python.framework.trading_env.broker_config import BrokerConfig
-        from python.framework.testing.mock_adapter import MockBrokerAdapter
+        from python.framework.testing.mock_broker_adapter import MockBrokerAdapter
         from python.framework.types.trading_env_types.broker_types import BrokerType
         from python.framework.logging.global_logger import GlobalLogger
