+++ tests/autotrader/live_executor/test_live_executor_modify.py
@@ -11,7 +11,7 @@ Tests modify_limit_order() via broker adapter:
 Uses DELAYED_FILL mode so orders stay in pending (broker_ref tracked).
 """
 
-from python.framework.testing.mock_adapter import MockBrokerAdapter, MockExecutionMode
+from python.framework.testing.mock_broker_adapter import MockBrokerAdapter, MockExecutionMode
 from python.framework.testing.mock_order_execution import MockOrderExecution
 from python.framework.trading_env.live.live_trade_executor import LiveTradeExecutor
 from python.framework.types.trading_env_types.order_types import (
