+++ tests/parity/test_trade_records_parity.py
@@ -19,7 +19,7 @@ Per #326 §10 Sim/Live Parity.
 from datetime import datetime, timezone
 
 from python.framework.logging.global_logger import GlobalLogger
-from python.framework.testing.mock_adapter import MockBrokerAdapter, MockExecutionMode
+from python.framework.testing.mock_broker_adapter import MockBrokerAdapter, MockExecutionMode
 from python.framework.testing.mock_order_execution import MockOrderExecution
 from python.framework.trading_env.broker_config import BrokerConfig
 from python.framework.trading_env.live.live_trade_executor import LiveTradeExecutor
