+++ tests/simulation/trade_emission/conftest.py
@@ -11,7 +11,7 @@ from datetime import datetime, timezone
 import pytest
 
 from python.framework.logging.global_logger import GlobalLogger
-from python.framework.testing.mock_adapter import MockBrokerAdapter, MockExecutionMode
+from python.framework.testing.mock_broker_adapter import MockBrokerAdapter, MockExecutionMode
 from python.framework.trading_env.broker_config import BrokerConfig
 from python.framework.trading_env.simulation.trade_simulator import TradeSimulator
 from python.framework.types.market_types.market_data_types import TickData
